+++ src/interface/TWData+Hex.cpp
@@ -40,7 +40,7 @@ TWData *TWDataCreateWithHexString(const TWString *hex) {
         uint8_t low = value(TWStringGet(hex, stringIndex));
         stringIndex += 1;
 
-        TWDataSet(data, dataIndex, (high << 4) | low);
+        TWDataSet(data, dataIndex, static_cast<uint8_t>((high << 4) | low));
         dataIndex += 1;
     }
